requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
python-magic==0.4.15
requests==2.18.4
urllib3==1.22
